+++ crates/nostr/src/prelude.rs
@@ -49,6 +49,7 @@ pub use crate::nips::nip21::{self, *};
 pub use crate::nips::nip26::{self, *};
 pub use crate::nips::nip34::{self, *};
 pub use crate::nips::nip35::{self, *};
+pub use crate::nips::nip38::{self, *};
 pub use crate::nips::nip39::{self, *};
 #[cfg(feature = "nip44")]
 pub use crate::nips::nip44::{self, *};