+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 
 ### Added
 
+* nostr: add NIP-38 support ([reyamir])
 * nostr: add NIP-62 support ([awiteb])
 * nostr: add `nip21::extract_from_text` function ([Yuki Kishimoto])
 * nostr: add `EventBuilder::allow_self_tagging` ([Yuki Kishimoto])